move type to types
